Form input reset + Input formatting + Style set
If the modal is closed the local state will reset to empty. This ensures
that in certain edge cases form data is still reset.

Form input is now formatted through a series of RegEx-s designed to
filter out any undesirable characters.

Inline style overrides which will probably be replaced with .css at a
later date