Added basic yaml processer (deposit)
At time of writing serde_yaml does not provide an adequate presentation
layer (as found in the 1.2 spec) thus I have added a C++ child process
written with cpp-yaml (credit to jbeder) to handle the display and
presentation. At this time it is incomplete and will fail if it detects
a raw scalar, however maps and sequences are tentatively working.