Consolidated and refactored config integration
settings.rs
- Finally removed my bastard singleton in favor of a simple rocket managed
struct. Most of the other changes listed below are a result of the
refactoring this required.
- Added clap support for modification of these values. As they are
required for the program to function they also have default values.
These will likely change in the future but for now they are set for my
dev environment.

routes.rs
- Added state handling parameters to the routes, rewrote the code that
relied on the singleton to instead make use of the new state.

models.rs
- Removed outdated comments
- Removed any singleton related code
- Simplified my path resolution

main.rs
- Added an AdHoc fairing to grab the extra config that rocket now must
manage